Refactoring in my branch models with new structure
Models Partners, Principles and Requisites
* Make new models files in models directory
* Make new admins files to models in admin directory
* Some pep8 fix
