+++ backend/npo_project/admin/__init__.py
@@ -1,6 +1,9 @@
-from .program_admin import ProgramAdmin, ProgramCharacteristicAdmin  # noqa
 from .course_price_admin import (  # noqa
     IncludedInCoursePriceAdmin,
     NotIncludedInCoursePriceAdmin,
     CoursePriceAdmin
 )
+from .partners_admin import PartnersAdmin  # noqa
+from .principles_admin import PrinciplesAdmin  # noqa
+from .program_admin import ProgramAdmin, ProgramCharacteristicAdmin  # noqa
+from .requisites_admin import RequisitesAdmin  # noqa
