+++ backend/npo_project/admin/requisites_admin.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+
+from ..models import Requisites
+
+
+@admin.register(Requisites)
+class RequisitesAdmin(admin.ModelAdmin):
+    list_display = (
+        'id',
+        'text',
+    )
+    list_editable = ('text',)
