+++ backend/npo_project/models/__init__.py
@@ -1,4 +1,7 @@
-from .program import Program, ProgramCharacteristic  # noqa
 from .course_price import (IncludedInCoursePrice,  # noqa
                            NotIncludedInCoursePrice,
                            CoursePrice)
+from .partners import Partners  # noqa
+from .principles import Principles  # noqa
+from .program import Program, ProgramCharacteristic  # noqa
+from .requisites import Requisites  # noqa
